Rediseño menu, funcion agregar item en materiales

+++ app/src/main/java/co/dolmen/sid/entidad/MovimientoArticulo.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class MovimientoArticulo implements Serializable {
     private int id_articulo;
     private int id_tipo_stock;
-    private double cantidad;
+    private float cantidad;
     private String articulo;
     private String tipo_stock;
     private String movimiento;
@@ -14,7 +14,7 @@ public class MovimientoArticulo implements Serializable {
 
     }
 
-    public MovimientoArticulo(int id_articulo, int id_tipo_stock, double cantidad, String articulo, String tipo_stock, String movimiento) {
+    public MovimientoArticulo(int id_articulo, int id_tipo_stock, float cantidad, String articulo, String tipo_stock, String movimiento) {
         this.id_articulo = id_articulo;
         this.id_tipo_stock = id_tipo_stock;
         this.cantidad = cantidad;
@@ -47,7 +47,7 @@ public class MovimientoArticulo implements Serializable {
         return movimiento;
     }
 
-    public void setCantidad(double cantidad) {
+    public void setCantidad(float cantidad) {
         this.cantidad = cantidad;
     }
 }
